packages/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Can Burak Sofyalıoğlu", email = "cbsofyalioglu@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "pydantic-ai>=0.0.18",
]

[project.scripts]
agents = "agents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
